Improve error handling for `/runfolders`
There is no reason why this handlers would return 404. If no runfolders
are found, it should just return an empty list. If another exception
pops up, it will be for a different reason and should be converted to an
Internal Error (500) (which aiohttp does automatically).

Same goes for `/runfolders/pickup` and `/runfolders/next`